Rename file and setup model and engine evaluation
