+++ evaluation/model_evaluation.py
@@ -1,0 +1,9 @@
+from dataset import EvaluationDataset
+
+class ModelEvaluator:
+    def __init__(self, dataset: EvaluationDataset, config={}):
+        self.dataset = dataset
+        self.config = config
+
+    def evaluate(self):
+        pass
